many kernel <> renderer interaction fixes
