+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/ChatWidgetHUD/ChatHUD.asmdef
@@ -3,12 +3,9 @@
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f",
         "GUID:b1087c5731ff68448a0a9c625bb7e52d",
-        "GUID:1de3566cccb280f4a982c59ad0d08c96",
         "GUID:bd51ff2b035ce4ab4be7c38c43c88889",
-        "GUID:858f845923fcab3448cf8b66614b41a1",
         "GUID:555c1f3c6d18648df910b7a1de75b424",
-        "GUID:eb95615676ca6984687c5ae7774d6309",
-        "GUID:4973650d2444c4561a15d50f24d91cd9"
+        "GUID:eb95615676ca6984687c5ae7774d6309"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
