+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/ChatWidgetHUD/ChatHUDController.cs
@@ -1,4 +1,3 @@
-using DCL.Interface;
 using System.Collections.Generic;
 using System.Linq;
 using ChatMessage = ChatController.ChatMessage;
@@ -8,11 +7,6 @@ public class ChatHUDController
     const int MAX_CHAT_ENTRIES = 100;
 
     public ChatHUDView view;
-
-    public void SendChatMessage(ChatMessage message)
-    {
-        WebInterface.SendChatMessage(message);
-    }
 
     public void AddChatMessage(ChatMessage message)
     {
