+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/WorldChatWindowHUD.asmdef
@@ -6,7 +6,9 @@
         "GUID:bd51ff2b035ce4ab4be7c38c43c88889",
         "GUID:858f845923fcab3448cf8b66614b41a1",
         "GUID:6c83ec3e64bad1743a40f7646cd964a4",
-        "GUID:eb95615676ca6984687c5ae7774d6309"
+        "GUID:eb95615676ca6984687c5ae7774d6309",
+        "GUID:1de3566cccb280f4a982c59ad0d08c96",
+        "GUID:4973650d2444c4561a15d50f24d91cd9"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
