+++ unity-client/Assets/Scripts/MainScripts/Debugging/WSS/WSSController.asmdef
@@ -11,7 +11,8 @@
         "GUID:31d8dc012a18a4a6da741a0fd9c390e0",
         "GUID:1a483cdd86db6e641ad77c7046dd035f",
         "GUID:8e14059e964ca49aeb55d0e7b2b14c3c",
-        "GUID:7240b5b323e5f4bde8392d35668ad0ae"
+        "GUID:7240b5b323e5f4bde8392d35668ad0ae",
+        "GUID:eb95615676ca6984687c5ae7774d6309"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
